requirement.txt
streamlit 
ollama
PyPDF2
faiss-cpu
python-dotenv
langchain_google_genai
langchain_community
pypdf